CR extracting detailConn code to a separate file
